requirements.txt
PySide6>=6.6
mss>=9.0
Pillow>=10.0
requests>=2.31
urllib3>=2.0

# Стабільність на Windows: OpenCV 4.12+ тягне numpy>=2, а багато пакетів досі чутливі.
numpy>=1.26,<2.0
opencv-python>=4.8,<4.12

# Локальний OCR (офлайн), щоб не мучити OCR.space
rapidocr>=3.3,<4.0.0
onnxruntime>=1.16


# Spellcheck (опційно, для виправлення OCR перед перекладом)
pyspellchecker>=0.8

# Argos Translate (офлайн переклад; мовні пакети ставляться окремо)
argostranslate>=1.9
